changes to map being used as base _low mem solution

+++ rtk/ClStr2Mat.h
@@ -49,6 +49,6 @@ private:
 	vector<string> smplLoc;
 	vector<string> baseP;
 	size_t smplN;
-	int curr ;
+	size_t curr ;
 };
